Extract Node and TOTP helpers
